+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add new `--dark` flag to enable rendering in dark mode.
+- Add new `--input-path`/`-i` option to use local copy of the swift-book repo.
+- Add a new detection mechanism for the minted requiring shell escape by typesetting a simple document containing a minted box.
+
+### Fixed
+- Fix an issue where some 2024 TeX Live installations may not be able to typeset TSPL because minted still requires the `-shell-escape` option.
 
 ### Internal
 - Change regular expression used for detecting URLs in Markdown that may have caused exponential backtracking on strings starting with '[\](http://' and containing many repetitions of '!'.